#105 handle response with null s3objects

+++ FluentStorage.AWS/Blobs/AwsS3DirectoryBrowser.cs
@@ -97,6 +97,9 @@ namespace FluentStorage.AWS.Blobs {
 			while (true) {
 				ListObjectsV2Response response = await _client.ListObjectsV2Async(request, cancellationToken).ConfigureAwait(false);
 
+				if (response?.S3Objects == null)
+					break;
+
 				await Task.WhenAll(response.S3Objects.Select(s3 => _client.DeleteObjectAsync(_bucketName, s3.Key, cancellationToken))).ConfigureAwait(false);
 
 				if (response.NextContinuationToken == null)
